Proyecto final/quiz-service/requirements.txt
flask
flask_cors
pymongo
psycopg2-binary
